Added PluginState static class that handles persistent configs
The code was extracted from GUIWindowBase. Now everything that needs to persist
its configs have to use PluginState.

For that there are two attributes: ConfigOption for fields and PersistState for
classes. The former marks fields that should be saved and restored, the later is
used to recursively save/load state of a tree of objects.