Fixed inability of the VesselSpawner to disable launched vessel colliders on time
1. Reimplemented VesselSpawner as a MonoBehaviour to be able to use FixedUpdate,
as WaitForFixedUpdate is executed AFTER the fixed update cycle, and many things
are already happened.
2. Using onVesselLoaded/OffRails events to invoke corresponding callbacks
instead of calling them at WaitForFixedUpdate stage.
3. The main methods are no longer coroutines. They launch corresponding
coroutines under the hood, as well as add some code to FixedUpdate. All so that
the VesselSpawner could control all aspect of the launch process.
4. To wait for the launch completion the VesselLauncher provides a
CustomYieldInstruction that is returned by the VesselLauncher.WaitForLaunch
property.

Fixes allista/hangar#213