Fixed jumps of dumped packed vessels out of the damper on vessel switch
The reason being in the transit between physical damping of rigid bodies and
kinematic damping of packed vessels. Namely, when a vessel was being packed it
skipped the frame before activating the next OnTriggerStay, and was carried away
by orbit adjustments.
Now, as soon as a damped part is packed, its vessel is added to the
dampedVessels, which allows to begin kinematic damping immediately.
The big change that was required for that to work, though, is that kinematic
damping should be done OnWaitForFixedUpdate, not in FixedUpdate (as we know from
VesselSpawner development).
So the damping of packed vessels had to be moved to a coroutine. Which brought
another problem -- at the end of a fixed frame positions are already updated; so
if we save relative position in FixedUpdate, it cannot be used to set position
AFTER this fixed frame has ended. So the VesselInfo is now created without the
position pre-set, and with the flag positionSet=false that indicates that the
damping coroutine sould calculate relative position at the first opportunity.

Also, host RB of the damper now receives its dL at the position of the damped
RB, which should result in more physically realistic behavior.