Refactored HSVPicker and ColorList/Item code; and fixed BoxSlider
BoxSlider threw several exceptions due to lame handling of shader-vs-texture
logic.

Also, changed the anchor of the ColorList window so that it could be placed at
at screen center (at least vertically).

Added ColorPicker prefab to the bundle and tuned it a bit.
Removed the GenerateSVTexture shader from the bundle, since it cannot be loaded
anyway.